The tracking sensor is programmed.
Right now the tracking sensor is lights up an led and nothing else. Also some deffinitions where changed. Left and right bottom sensor are not fixed and therefore in the deffinition it is the same pin as Middle Line tracking sensor. That is no problem because these sensor are not initialized.